+++ src/index.test.ts
@@ -109,11 +109,14 @@ import {
   STEP_CLOUD_RUN_ROUTES,
   STEP_CLOUD_RUN_SERVICES,
 } from './steps/cloud-run/constants';
-import { STEP_REDIS_INSTANCES } from './steps/redis/constants';
 import {
   STEP_CREATE_MEMCACHE_INSTANCE_NETWORK_RELATIONSHIPS,
   STEP_MEMCACHE_INSTANCES,
 } from './steps/memcache/constants';
+import {
+  STEP_CREATE_REDIS_INSTANCE_NETWORK_RELATIONSHIPS,
+  STEP_REDIS_INSTANCES,
+} from './steps/redis/constants';
 import {
   STEP_SPANNER_INSTANCES,
   STEP_SPANNER_INSTANCE_CONFIGS,
@@ -634,6 +637,9 @@ describe('#getStepStartStates success', () => {
         disabled: false,
       },
       [STEP_REDIS_INSTANCES]: {
+        disabled: false,
+      },
+      [STEP_CREATE_REDIS_INSTANCE_NETWORK_RELATIONSHIPS]: {
         disabled: false,
       },
       [STEP_MEMCACHE_INSTANCES]: {