+++ src/steps/redis/constants.ts
@@ -1,4 +1,6 @@
 export const STEP_REDIS_INSTANCES = 'fetch-redis-instances';
+export const STEP_CREATE_REDIS_INSTANCE_NETWORK_RELATIONSHIPS =
+  'build-redis-instance-network-relationships';
 
 export const ENTITY_TYPE_REDIS_INSTANCE = 'google_redis_instance';
 export const ENTITY_CLASS_REDIS_INSTANCE = ['Database', 'DataStore', 'Host'];